Improve corners and button using CSS3. Remove jQuery.
It shows as intended in Opera 10.5x, then degrade gracefully in
Chrome > Firefox > IE. Safari is probably around the same as Chrome.
Note that Firefox 3.6.x SHOULD work but doesn't for me.

Either way it's better this way as we remove Javascript and we won't
have to update the source to take care of newer browser support.